Major NuShell Adjustment + Starship
I've made major changes to NuShell to ensure that carapace is working properly, it now points to this file regardless of what windows pc I'm on, and it has a starship prompt setup! To achieve the right fonts for the starship, I've had to make adjustments to the font family within alacritty, so that's also been adjusted.