Complete GlazeWM Revamp
Heavy reformatting on the yaml, ensuring that it doesn't have irrelevant code (such as code I can always just look back on from within git history), and that it's updated for what I need now. This has completely changed it. I also created a symlink from the glaze config default location to here, so that now can be called at startup with just the exe, and every time it uses the exe it refers to the config here! Some changes will needed to be uncommented on the desktop pc.